requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.24
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.3
passlib[bcrypt]==1.7.4
